requirements.txt
pandas
ddgs
groq
python-dotenv
